backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
loguru>=0.7.0
psycopg2-binary
python-jose
minio>=7.1.0